+++ packages/block-library/src/paragraph/edit.js
@@ -6,11 +6,14 @@ import classnames from 'classnames';
 /**
  * WordPress dependencies
  */
+import { useState } from '@wordpress/element';
 import { __, _x, isRTL } from '@wordpress/i18n';
 import {
 	ToolbarButton,
 	ToggleControl,
 	__experimentalToolsPanelItem as ToolsPanelItem,
+	DropZone,
+	Popover,
 } from '@wordpress/components';
 import {
 	AlignmentControl,
@@ -19,9 +22,13 @@ import {
 	RichText,
 	useBlockProps,
 	useSetting,
+	store as blockEditorStore,
 } from '@wordpress/block-editor';
+import { useDispatch } from '@wordpress/data';
+import { useMergeRefs } from '@wordpress/compose';
 import { createBlock } from '@wordpress/blocks';
 import { formatLtr } from '@wordpress/icons';
+import { createBlobURL } from '@wordpress/blob';
 
 /**
  * Internal dependencies
@@ -55,14 +62,22 @@ function ParagraphBlock( {
 } ) {
 	const { align, content, direction, dropCap, placeholder } = attributes;
 	const isDropCapFeatureEnabled = useSetting( 'typography.dropCap' );
+	const [ paragraphElement, setParagraphElement ] = useState( null );
+	const refCallback = ( element ) => {
+		setParagraphElement( element );
+	};
 	const blockProps = useBlockProps( {
-		ref: useOnEnter( { clientId, content } ),
+		ref: useMergeRefs( [
+			useOnEnter( { clientId, content } ),
+			refCallback,
+		] ),
 		className: classnames( {
 			'has-drop-cap': dropCap,
 			[ `has-text-align-${ align }` ]: align,
 		} ),
 		style: { direction },
 	} );
+	const { replaceBlock } = useDispatch( blockEditorStore );
 
 	return (
 		<>
@@ -107,6 +122,36 @@ function ParagraphBlock( {
 						/>
 					</ToolsPanelItem>
 				</InspectorControls>
+			) }
+			{ ! content && (
+				<Popover
+					anchorRef={ paragraphElement }
+					animate={ false }
+					position="top right left"
+					focusOnMount={ false }
+					__unstableForcePosition
+					className="wp-block-paragraph__drop_zone"
+				>
+					<DropZone
+						style={ {
+							// TODO: Ideally we should observe the size of the paragraph block.
+							width: paragraphElement?.offsetWidth,
+							height: paragraphElement?.offsetHeight,
+						} }
+						onFilesDrop={ ( files ) => {
+							if ( files.length === 1 ) {
+								replaceBlock(
+									clientId,
+									createBlock( 'core/image', {
+										url: createBlobURL( files[ 0 ] ),
+									} )
+								);
+							}
+
+							// TODO: We can handle other file types and sizes here.
+						} }
+					/>
+				</Popover>
 			) }
 			<RichText
 				identifier="content"