+++ packages/block-library/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### New Feature
 
 -   Made it possible to import individual blocks ([#42258](https://github.com/WordPress/gutenberg/pull/42258)). Check [README](./README.md#loading-individual-blocks) for more information.
+-   Paragraph block: You can now drag an image on an empty Paragraph block to create an image block ([#42722](https://github.com/WordPress/gutenberg/pull/42722)).
 
 ## 7.13.0 (2022-08-24)